requirements.txt
transformers
torch
torchaudio
sentencepiece
language-tool-python
openai-whisper
ffmpeg-python
deepl
langdetect
pyspellchecker
requests
